added an unstable algrithm and average reward tests
